refactor ExecuteWriteTask
